improving AI algorithm representation
