+++ artificial_intellgence/README.md
@@ -1,0 +1,22 @@
+# Human decision-making by machines
+
+Enabling a machine to make a decision by mimicing a human behavior is something that can be corresponded to Artificial Intelligence (AI). The big picture of AI is shown in the picture below.
+
+![AI, ML, and DL](images/1280px-AI-ML-DL.svg.png)
+
+### Supervised Learning
+
+Supervised learning is a machine learning paradigm for problems where the available data consists of labelled examples, meaning that each data point contains features (covariates) and an associated label. The goal of supervised learning algorithms is learning a function that maps feature vectors (inputs) to labels (output), based on example input-output pairs.
+
+### Semi-Supervised Learning
+
+Semi-supervised learning is an approach to machine learning that combines a small amount of labeled data with a large amount of unlabeled data during training. Semi-supervised learning falls between unsupervised learning (with no labeled training data) and supervised learning (with only labeled training data). It is a special instance of weak supervision.
+
+### Unsupervised Learning
+
+Unsupervised learning is a machine learning paradigm for problems where the available data consists of unlabelled examples, meaning that each data point contains features (covariates) only, without an associated label. The goal of unsupervised learning algorithms is learning useful patterns or structural properties of the data. Examples of unsupervised learning tasks are clustering, dimension reduction, and density estimation.
+
+* <https://en.wikipedia.org/wiki/Deep_learning>
+* <https://en.wikipedia.org/wiki/Semi-supervised_learning>
+* <https://en.wikipedia.org/wiki/Supervised_learning>
+* <https://en.wikipedia.org/wiki/Unsupervised_learning>
