+++ artificial_intellgence/supervised_learning/neural_networks/neural_network/convolution_neural_network.py
