+++ artificial_intellgence/supervised_learning/neural_networks/neural_network/gan.py_tf
