--- neural_network/deep_learning/README.md
@@ -1,5 +1,0 @@
-# Deep Learning
-
-Deep learning is a class of machine learning algorithms that uses multiple layers to progressively extract higher-level features from the raw input. For example, in image processing, lower layers may identify edges, while higher layers may identify the concepts relevant to a human such as digits or letters or faces.
-
-* <https://en.wikipedia.org/wiki/Deep_learning#:~:text=Deep%20learning%20is%20a%20class,digits%20or%20letters%20or%20faces.>
