--- neural_network/deep_learning/reinforcement_learning/README.md
@@ -1,5 +1,0 @@
-# Reinforcement Learning
-
-Reinforcement learning (RL) is an area of machine learning concerned with how intelligent agents ought to take actions in an environment in order to maximize the notion of cumulative reward. Reinforcement learning is one of three basic machine learning paradigms, alongside supervised learning and unsupervised learning.
-
-* <https://en.wikipedia.org/wiki/Reinforcement_learning>
