--- neural_network/deep_learning/supervised_learning/README.md
@@ -1,5 +1,0 @@
-# Supervised Learning
-
-Supervised learning (SL) is a machine learning paradigm for problems where the available data consists of labelled examples, meaning that each data point contains features (covariates) and an associated label. The goal of supervised learning algorithms is learning a function that maps feature vectors (inputs) to labels (output), based on example input-output pairs.
-
-* <https://en.wikipedia.org/wiki/Supervised_learning>
