--- neural_network/deep_learning/unsupervised_learning/README.md
@@ -1,5 +1,0 @@
-# Unsupervised Learning
-
-Unsupervised learning is a machine learning paradigm for problems where the available data consists of unlabelled examples, meaning that each data point contains features (covariates) only, without an associated label. The goal of unsupervised learning algorithms is learning useful patterns or structural properties of the data. Examples of unsupervised learning tasks are clustering, dimension reduction, and density estimation.
-
-* <https://en.wikipedia.org/wiki/Unsupervised_learning>
